Include additional role service methods for example authorization
Add methods and a few corrections to allow galleries (via the
canvasGallery.js example) to be shared to other users.  The UI is still
in progress, but the service hooks are available so an authorized user
can a) create a new personal role, b) authorize that role to one or
more groups or data items, and c) add another user to that role.  At
the moment, access to the user list exposes contact info, so this needs
to be limited and/or the authorization mechanism changed.