Add persistence option for protected credential
This credential is used to protect the private key, and is itself
enciphered with a dedicated key stored in the database.  This
arrangement allows for the private key to be protected, and that
protection to be stored elsewhere from the database and the private key,
but in a portable format.