Add KBV Demo
This is an example of how a Knowledge-based Verification might be done.
It does not include anything about the UI or workflow, only the policy
execution itself to compare supplied values.

NOTE 1: Nobody wants this type of thing anymore, yet a lot of people
tend to need this.  So it's counterintuitive to include it, but I did so
to see if I could.

NOTE 2: I stopped short of completing this because a) I don't like the
way it's currently arranged, and 2) the goal should be not to need this
type of thing, ever