Tweak/Phase-shift the session change announce versus the direct operation invocation
The whole operation/rule paradigm was largely experimental with the
original 5.x version, but is somewhat cumbersome to use and I never
wound up liking it very much.  In this case it created a problem where
the session change announce handler instigates one action, while the
operation instigates another, and they both collide.