Basic peek and variable use operations
Add in a basic peek and variable cache operation.  The peek
(tab-to-complete) is rudimentary but serviceable.  The variable cache
takes any matching object from the current group or data list and
stashes the reference in a variable array ($#) that can then be
referenced for basic operations.  One all the client model references
are consolidated, this should allow for greater flexibility in directly
working with the objects via a psuedo console interface versus having to
script around the various dependencies and API calls.  Look at the
policy.js implementation for an early example of what I'm thinking.