Add initial plumbing for approver evaluation
Add initial plumbing for evaluating approvals.  The methodology draws on the policy and control mechanisms, whereby a control may be applied to an entitlement (ROLE, PERMISSION, or GROUP) , and associated with a policy.  The policy then defines patterns of type APPROVAL, and the cited facts leverage the corresponding operations to extract the associated approvers.  This is largely in-flight, with Request and Approval factories and evaluation to follow.