Add application generator UX
Add UX to generate an application definition.  When in community mode with a community project selected, this create accounts for all community project scoped persons, add a random assortment of groups and/or permissions, and randomly assign those to the accounts.

This is for testing access traceability and performance in IAM ecosystems.