Moved enablement into a separate method
Most of the object UXs have same basic logic to limit applicability so they aren't randomly launched and throw errors for incompatible types.  However, for the tag picker, this can be more indescriminate so this allows the reference object to make sure the UI is enable regardless of any matching object type.