Add check for null table
This is for a very particular use case where the data access library is started and unceremoniously stopped by the console app when performing operations that don't involve the DAL.  While it should be failing gracefully, or otherwise logging the error and trace, it wasn't and throwing an NPE.  Use case is: Start Factories, don't use factories, shut down factories and threads.  There must be in missed initialization here.