Fix bug with opening a profile for a user without a person reference
There was a bug when opening the profile view for a user without a person object, where an object being passed may or may not have been a promise, and if a promised person, may or may not exist.  If a promised person doesn't exist, then switch back to the user as the preferred profile object.