Fixed bug when patching data objects
Fixed a bug where an attempt is made to patch a DataType object meta data when it was read as a full object (the bytestore is removed prior to patching), and the compression bit was being incorrectly set, thereby munging the byte array on the subsequent read operation.  The correction was to move the update to the compression bit and type to happen along with the bytestore, versus allowing it to be updated with the other metadata attribute patching.