Incremental update to testing extraction of linked identity entitlements
This test covers pulling out all of the entitlements for a person, user, or account, and the subsequent linked values.  For example: A user linked to multiple persons, or a person linked to a user and multiple accounts, so that the person or user reference will include entitlements for the linked objects.