Add additional meta data, and enclose content HTML
The section content HTML is enclosed to optionally inject adjacent content when using the example 'picture mode' (tag data with a tag named 'picture mode') which will reorganize the article into a gallery of sorts.  This could be better handled with CSS, but for the moment I was just trying to see if it would work.