Fix issue with deserializing ArrayList<?>
Jackson was returning the generic type as <LinkedHashMap> for complex objects.  Temporarily went with the fix of using an ObjectMapper against the raw JSON. When using the ObjectMapper, passing in List<SomeType> still resulted in the LinkedHashMap, and trying to return List<SomeType> off the generic class still returned LinkedHashMap.  So, using a secondary method only for lists seemed to resolve this.  There's likely a better solution here.