Add service to count and search for entitlements and objects authorized to the specified actor
This is a direct-to-data query versus the layers of authorization used through the BaseService, which is useful for passing objects back and forth.  However, when wanting to search outside of a group or parent container, that no longer works.  This query, an adaption from the version for looking only for entitlements, will accept search criteria, pagination instructions, and can return either the EntitlementType or the actual object type for the authorized actor.