Add in basic authorization support for spool types (message specifically)
The spool is a unique construct in that it is meant to capture variable state entries and was not originally designed for direct user interaction, usually masked behind other entities such as security tokens and access requests.  This is only a very coarse authorization check, and ideally should follow the main authorization evaluation chain.